pyproject.toml
[tool.poetry]
name = "injecta"
version = "0.9.3"
description = "Dependency Injection Container Library"
readme = "README.md"
repository = "https://github.com/pyfony/injecta"

license = "MIT"

authors = [
    "Jiri Koutny <jiri.koutny@datasentics.com>"
]

classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

packages = [
    { include = "injecta", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.7"
python-box = "^3.4"
PyYAML = "^5.1"
tomlkit = ">=0.5.8,<1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^20.8b1"
flake8 = "^3.8.4"
poethepoet = "^0.10.0"
pep8-naming = "^0.11.1"
flake8-fixme = "^1.1.1"

[tool.black]
line-length = 140

[tool.poe.tasks]
black = "black src"
black-check = "black --check src"
flake8 = "flake8 --ignore=E501,W503 src"